Proper type used

+++ modules/type/blocks/dap_chain_cs_blocks.c
@@ -363,7 +363,7 @@ static int s_cli_blocks(int a_argc, char ** a_argv, char **a_str_reply)
     };
     const size_t l_subcmd_str_count=sizeof(l_subcmd_strs)/sizeof(*l_subcmd_strs)-1;
     const char* l_subcmd_str_args[l_subcmd_str_count];
-	for(int i=0;i<l_subcmd_str_count;i++)
+	for(size_t i=0;i<l_subcmd_str_count;i++)
         l_subcmd_str_args[i]=NULL;
     const char* l_subcmd_str_arg;
     const char* l_subcmd_str = NULL;
